+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 ## 0.0.3
 
 - Add few styles to the widget for better look
+
+## 0.0.4
+
+- Changed PrefixText to be a prefix Widget
+- Added a suffix Widget
+- Added disableTextFieldStyle to disable the text field style
+- Added textFormFieldStyle to customize the text field style
+- Added a blog post link to the README